src/main/java/com/realcomp/data/conversion/ConstantConverter.java
package com.realcomp.data.conversion;

/**
 *
 * @author krenfro
 */
@com.realcomp.data.annotation.Converter("constant")
public class ConstantConverter implements Converter{

    protected String value = "";

    @Override
    public String convert(String value) throws ConversionException{

        if (value == null)
            throw new IllegalArgumentException("value is null");

        return this.value;
    }

    public String getValue() {
        return value;
    }

    public void setValue(String value) {
        this.value = value;
    }

    @Override
    public boolean equals(Object obj) {
        if (obj == null)
            return false;
        if (getClass() != obj.getClass())
            return false;
        final ConstantConverter other = (ConstantConverter) obj;
        if ((this.value == null) ? (other.value != null) : !this.value.equals(other.value))
            return false;
        return true;
    }

    @Override
    public int hashCode() {
        int hash = 3;
        hash = 97 * hash + (this.value != null ? this.value.hashCode() : 0);
        return hash;
    }
}
